fix: implement missing TUI types to fix build and tests
Agent-Logs-Url: https://github.com/edwifiguy/OpenMonoAgent.ai/sessions/293ceab3-54d6-41a4-9ba5-660ed4ada8c4

+++ src/OpenMono.Cli/Tui/ContextWindowMeter.cs
@@ -11,7 +11,7 @@ public class ContextWindowMeter
 
     public int PromptTokens { get; private set; }
 
-    public int UsagePercent => _contextSize > 0 ? PromptTokens * 100 / _contextSize : 0;
+    public int UsagePercent => _contextSize > 0 ? (int)((long)PromptTokens * 100 / _contextSize) : 0;
 
     public int RemainingTokens => Math.Max(0, _contextSize - PromptTokens);
